Add a __version__ and version_info attribute.
This make it slightly easier to dynamically check version.

The __version__ one is relatively classic.
The version_info a bit less, but is s tuple which is easier to compare,
especially in Python 3 as you do not need to take a slice, without it
people usually will start to compare the string, which end up failing
if you ever change one of the version fields to be two digits.